feat: Update installation scripts to support macOS and improve package management

+++ dotfiles/git/install.sh
@@ -2,10 +2,14 @@
 
 set -e
 
-# Install/update lazy git
-LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
-curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
-tar xf lazygit.tar.gz lazygit
-sudo install lazygit /usr/local/bin
-rm lazygit.tar.gz lazygit
+# Install/update lazy git - skip on macOS (Darwin)
+if [[ "$OSTYPE" != "darwin"* ]]; then
+	LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
+	curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
+	tar xf lazygit.tar.gz lazygit
+	sudo install lazygit /usr/local/bin
+	rm lazygit.tar.gz lazygit
+else
+	echo "Skipping lazy git installation on macOS (Darwin)"
+fi
 # lazygit --version
